tests/wpt/web-platform-tests/permissions-policy/reporting/usb-report-only.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script src='/resources/testdriver.js'></script>
    <script src="/resources/testdriver-vendor.js"></script>
  </head>
  <body>
    <div id='fs'></div>
    <script>
var check_report_format = ([reports, observer]) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.body.featureId, "usb");
  assert_equals(report.body.disposition, "report");
};

promise_test(async t => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['feature-policy-violation']}).observe();
  });
  await test_driver.bless('Activate document for USB');
  await navigator.usb.getDevices();
  check_report_format(await report);
}, "USB report only mode");
    </script>
  </body>
</html>
